\name{sessionTools-package}
\alias{sessionTools-package}
\alias{sessionTools}
\docType{package}
\title{
Utilities for Saving and Restoring Sessions
}
\description{
Contiains tools that can be used to save and restore R sessions. This includes
objects, search paths, options and installed libraries.
}
\details{
\tabular{ll}{
Package: \tab sessionTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-16\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matthew D. Furia <matt.furia@sagebase.org>

Maintainer: Matthew D. Furia <matt.furia@sagebase.org>
}

