\name{sessionTools-package}
\alias{sessionTools-package}
\alias{sessionTools}
\docType{package}
\title{
Utilities for Saving and Restoring Sessions
}
\description{
Contiains tools that can be used to save and restore R sessions. This includes
objects, search paths, options and installed libraries. The current version
installs missing packages, but does not restore the exact package version. This
functionality will be added in the future, if it is feasible to do so.
}
\details{
\tabular{ll}{
Package: \tab sessionTools\cr
Type: \tab Package\cr
Version: \tab 0.9-1\cr
Date: \tab 2012-02-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matthew D. Furia \email{matt.furia@sagebase.org}

\url{https://github.com/MattNapsAlot}

Maintainer: Matthew D. Furia \email{matt.furia@sagebase.org}
}

